Move the `file-store-local` module under the `@bayou` namespace.
