+++ local-modules/@bayou/file-store-local/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "file-store-local",
+  "name": "@bayou/file-store-local",
   "version": "1.0.0",
 
   "dependencies": {
